+++ packages/enhanced-img/src/preprocessor.js
@@ -3,9 +3,8 @@ import * as path from 'node:path';
 
 import MagicString from 'magic-string';
 import { asyncWalk } from 'estree-walker';
+import { VERSION } from 'svelte/compiler';
 import { parse } from 'svelte-parse-markup';
-
-const ASSET_PREFIX = '___ASSET___';
 
 // TODO: expose this in vite-imagetools rather than duplicating it
 const OPTIMIZABLE = /^[^?]+\.(avif|heif|gif|jpeg|jpg|png|tiff|webp)(\?.*)?$/;
@@ -35,10 +34,19 @@ export function image(opts) {
 			const s = new MagicString(content);
 			const ast = parse(content, { filename });
 
-			// Import path to import name
-			// e.g. ./foo.png => ___ASSET___0
-			/** @type {Map<string, string>} */
+			/**
+			 * Import path to import name
+			 * e.g. ./foo.png => __IMPORTED_ASSET_0__
+			 * @type {Map<string, string>}
+			 */
 			const imports = new Map();
+
+			/**
+			 * Vite name to declaration name
+			 * e.g. __VITE_ASSET_0__ => __DECLARED_ASSET_0__
+			 * @type {Map<string, string>}
+			 */
+			const consts = new Map();
 
 			/**
 			 * @param {import('svelte/types/compiler/interfaces').TemplateNode} node
@@ -94,10 +102,10 @@ export function image(opts) {
 						image = await process(resolved_id, opts);
 						images.set(resolved_id, image);
 					}
-					s.update(node.start, node.end, img_to_picture(content, node, image));
+					s.update(node.start, node.end, img_to_picture(consts, content, node, image));
 				} else {
-					// e.g. <img src="./foo.svg" /> => <img src={___ASSET___0} />
-					const name = ASSET_PREFIX + imports.size;
+					// e.g. <img src="./foo.svg" /> => <img src={__IMPORTED_ASSET_0__} />
+					const name = '__IMPORTED_ASSET_' + imports.size + '__';
 					const { start, end } = src_attribute;
 					// update src with reference to imported asset
 					s.update(
@@ -129,18 +137,27 @@ export function image(opts) {
 				}
 			});
 
-			// add imports
+			// add imports and consts to <script module> block
+			let text = '';
 			if (imports.size) {
-				let import_text = '';
 				for (const [path, import_name] of imports.entries()) {
-					import_text += `import ${import_name} from "${path}";`;
+					text += `\timport ${import_name} from "${path}";\n`;
 				}
-				if (ast.instance) {
-					// @ts-ignore
-					s.appendLeft(ast.instance.content.start, import_text);
-				} else {
-					s.append(`<script>${import_text}</script>`);
+			}
+
+			if (consts.size) {
+				for (const [vite_name, declaration_name] of consts.entries()) {
+					text += `\tconst ${declaration_name} = "${vite_name}";\n`;
 				}
+			}
+
+			if (ast.module) {
+				// @ts-ignore
+				s.appendLeft(ast.module.content.start, text);
+			} else {
+				s.prepend(
+					`<script ${VERSION.startsWith('4') ? 'context="module"' : 'module'}>${text}</script>\n`
+				);
 			}
 
 			return {
@@ -233,8 +250,8 @@ function serialize_img_attributes(content, attributes, details) {
 	/** @type {number | undefined} */
 	let user_height;
 	for (const attribute of attributes) {
-		if (attribute.name === 'width') user_width = parseInt(attribute.value[0]);
-		if (attribute.name === 'height') user_height = parseInt(attribute.value[0]);
+		if (attribute.name === 'width') user_width = parseInt(attribute.value[0].raw);
+		if (attribute.name === 'height') user_height = parseInt(attribute.value[0].raw);
 	}
 	if (!user_width && !user_height) {
 		attribute_strings.push(`width=${details.width}`);
@@ -264,11 +281,12 @@ function stringToNumber(param) {
 }
 
 /**
+ * @param {Map<string,string>} consts
  * @param {string} content
  * @param {import('svelte/types/compiler/interfaces').TemplateNode} node
  * @param {import('vite-imagetools').Picture} image
  */
-function img_to_picture(content, node, image) {
+function img_to_picture(consts, content, node, image) {
 	/** @type {Array<import('svelte/types/compiler/interfaces').BaseDirective | import('svelte/types/compiler/interfaces').Attribute | import('svelte/types/compiler/interfaces').SpreadAttribute>} attributes */
 	const attributes = node.attributes;
 	const index = attributes.findIndex((attribute) => attribute.name === 'sizes');
@@ -281,11 +299,11 @@ function img_to_picture(content, node, image) {
 	let res = '<picture>';
 
 	for (const [format, srcset] of Object.entries(image.sources)) {
-		res += `<source srcset=${to_value(srcset)}${sizes_string} type="image/${format}" />`;
+		res += `<source srcset=${to_value(consts, srcset)}${sizes_string} type="image/${format}" />`;
 	}
 
 	res += `<img ${serialize_img_attributes(content, attributes, {
-		src: to_value(image.img.src),
+		src: to_value(consts, image.img.src),
 		width: image.img.w,
 		height: image.img.h
 	})} />`;
@@ -294,10 +312,19 @@ function img_to_picture(content, node, image) {
 }
 
 /**
+ * @param {Map<string, string>} consts
  * @param {string} src
  */
-function to_value(src) {
-	return src.startsWith('__VITE_ASSET__') ? `{"${src}"}` : `"${src}"`;
+function to_value(consts, src) {
+	if (src.startsWith('__VITE_ASSET__')) {
+		let var_name = consts.get(src);
+		if (!var_name) {
+			var_name = '__DECLARED_ASSET_' + consts.size + '__';
+			consts.set(src, var_name);
+		}
+		return `{${var_name}}`;
+	}
+	return `"${src}"`;
 }
 
 /**